+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## [unreleased]
+
+### 🧪 Testing
+
+- *(core)* Streamline database availability check and enhance benchmark tests
+
 ## [1.55.4] - 2026-06-30
 
 ### 🐛 Bug Fixes